+++ libs/wallet-react/src/components/connect-dialog/connection-options/connector-icon.tsx
