+++ libs/wallet-react/src/components/connect-dialog/index.ts
@@ -1,8 +1,6 @@
 export { ConnectDialogWithRiskAck, ConnectDialog } from './connect-dialog';
-export {
-  ConnectionOptions,
-  ConnectionOptionDefault,
-  ConnectionOptionRecord,
-} from './connection-options';
+export { ConnectionOptions } from './connection-options';
 export { ConnectionStatus } from './connection-status';
-export { ConnectorIcon } from './connector-icon';
+export { ConnectorIcon } from './connection-options/connector-icon';
+export { ConnectionOptionDefault } from './connection-options/connection-option-default';
+export { ConnectionOptionRecord } from './connection-options/custom-connection-options';
